python/requirements.txt
pyserial==3.5
matplotlib>=3.8.0
numpy>=1.26.0
pandas>=2.0.0
openpyxl>=3.1.0
pyinstaller>=5.13.0 